intent: Widen to u16, use a repr enum, and put intent first
Addresses review feedback on the signing-domain registry:

- Intent is now a #[repr(u16)] enum with explicit discriminants instead
  of loose u8 consts, so the compiler rejects two message types sharing
  a value (mirrors the guardian IntentType pattern). Values use the u16
  space as per-domain blocks (core 0x00xx, bitcoin 0x01xx) with room to
  grow; future chains claim 0x02xx onward.
- The signing preimage leads with the intent: intent (u16 LE) ||
  bcs(epoch) || bcs(message), across the committee verify path, the
  proof-of-possession preimage, the off-chain signing helper, and the
  Move test cert-message helpers.
- The on-wire value is the discriminant via 'as u16', never serde's
  enum encoding (which is the variant index). Added intent_values_are_stable
  and a preimage-layout test to lock the format; e2e remains the true
  cross-language check.

+++ packages/hashi/sources/hashi.move
@@ -72,7 +72,7 @@ public(package) fun assert_unpaused(self: &Hashi) {
 /// Returns the certified message (message + signature + stake support).
 public(package) fun verify<T>(
     self: &Hashi,
-    intent: u8,
+    intent: u16,
     message: T,
     sig: CommitteeSignature,
 ): CertifiedMessage<T> {
@@ -86,7 +86,7 @@ public(package) fun verify<T>(
 public(package) fun verify_with_committee<T>(
     _self: &Hashi,
     committee: &Committee,
-    intent: u8,
+    intent: u16,
     message: T,
     sig: CommitteeSignature,
 ): CertifiedMessage<T> {
